Web/pyproject.toml
[project]
name = "webdemoleavesdisease"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "streamlit>=1.52.1",
    "tensorflow>=2.20.0",
]
